Move example to scala example
