--- apps/streaming-inference/streaming-text-classification/README.md
@@ -1,8 +1,0 @@
-# Streaming Text Classification
-Based on Streaming example NetworkWordCount and Zoo text classification example. Network inputs (Strings) are pre-processed and classified by zoo. We applied a simple text classification model based on zoo example.
-
-
-## 1. Local case
-
-
-## 2. Distrubted case
