+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/streaming/README.md
@@ -1,0 +1,19 @@
+# Streaming Inferece Examples
+Streaming inference/predict based on [analytics-zoo](https://github.com/intel-analytics/analytics-zoo)
+
+## Summary
+Quick example about integrating analytics-zoo inference/predict service into streaming related applications. We prepared 2 examples: [Streaming Object Detection]() and [Streaming Text Classification](). Streaming Object Detection uses pre-trained SSD model to detect objects in images, while Streaming Text Classification uses pre-trained LSTM model to classify text from network input.
+
+## Environment
+* Apache Spark 1.6.0/2.1.0 (This version needs to be same with the version you use to build Analytics Zoo)
+
+## Datasets and pre-trained models
+
+**1. Streaming Object Detection**
+* Datasets: [COCO](http://cocodataset.org/#home)
+* Pre-trained model: [SSD 300x300](https://s3-ap-southeast-1.amazonaws.com/analytics-zoo-models/object-detection/analytics-zoo_ssd-vgg16-300x300_COCO_0.1.0.model)
+
+**2. Streaming Text Classification**
+* Word embeding file: [golve](https://nlp.stanford.edu/projects/glove/)
+* Pre-trained model: Save trained text classification model and word index in [Text Classification](https://github.com/intel-analytics/analytics-zoo/blob/master/docs/docs/ProgrammingGuide/text-classification.md).
+
