+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/streaming/objectdetection/ImagePathWriter.scala
@@ -1,28 +1,6 @@
-/*
- * Copyright 2018 Analytics Zoo Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.intel.analytics.zoo.apps.streaming
+package com.intel.analytics.zoo.examples.streaming.objectdetection
 
 import java.io.{File, PrintWriter}
-
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.FileSystem
-import org.apache.hadoop.fs.Path
-import org.apache.log4j.{Level, Logger}
-import scopt.OptionParser
 
 /*
   * Based on https://github.com/intel-analytics/meph-streaming
