+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/streaming/textclassification/StreamingTextClassification.scala
@@ -14,23 +14,32 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.zoo.apps.streaming
-
-import com.intel.analytics.bigdl.tensor.Tensor
+package com.intel.analytics.zoo.examples.streaming.textclassification
 
 import scala.collection.immutable._
 import scala.io.Source
-import org.apache.spark.sql.functions._
 import com.intel.analytics.zoo.common.NNContext
 import com.intel.analytics.zoo.feature.text.{TextFeature, TextSet}
 import com.intel.analytics.zoo.models.textclassification.TextClassifier
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric.NumericFloat
-import org.apache.spark.sql.SparkSession
+import org.apache.spark.storage.StorageLevel
+import org.apache.spark.streaming.{Seconds, StreamingContext}
 import scopt.OptionParser
 
 
+case class TextClassificationParams(
+  host: String = "localhost",
+  port: Int = 9999,
+  indexPath: String = "textclassification.txt",
+  embeddingPath: String = "./",
+  classNum: Int = 20, tokenLength: Int = 200,
+  sequenceLength: Int = 500, encoder: String = "cnn",
+  encoderOutputDim: Int = 256, maxWordsNum: Int = 5000,
+  trainingSplit: Double = 0.8, batchSize: Int = 128,
+  nbEpoch: Int = 20, learningRate: Double = 0.01,
+  partitionNum: Int = 4, model: Option[String] = None)
 
-object StructuredStreamingTextClassification {
+object StreamingTextClassification {
 
   def main(args: Array[String]) {
     val parser = new OptionParser[TextClassificationParams]("TextClassification Example") {
@@ -76,55 +85,47 @@ object StructuredStreamingTextClassification {
         .action((x, c) => c.copy(model = Some(x)))
     }
 
-    parser.parse(args, TextClassificationParams()).map { param =>
-      val sc = NNContext.initNNContext("Structured Network Text Streaming Predict")
-      val spark = SparkSession
-        .builder.config(sc.getConf)
-        .getOrCreate()
 
-      import spark.implicits._
+
+    parser.parse(args, TextClassificationParams()).map { param =>
+      val sc = NNContext.initNNContext("Network Text Streaming Predict")
+      val ssc = new StreamingContext(sc, Seconds(3))
 
       val wordIndex: Map[String, Int] = readWordIndex(param.indexPath)
+
       val model = TextClassifier.loadModel[Float](param.model.get)
-      
-      val lines = spark.readStream
-        .format("socket")
-        .option("host", param.host)
-        .option("port", param.port)
-        .load()
+      // Create a socket stream on target ip:port and count the
+      // words in input stream of \n delimited text (eg. generated by 'nc')
+      // Note that no duplication in storage level only for running locally.
+      // Replication necessary in distributed scenario for fault tolerance.
+      val lines = ssc.socketTextStream(param.host,
+        param.port, StorageLevel.MEMORY_AND_DISK_SER)
 
-      // Predict UDF
-      val predictTextUDF = udf ( (x: String) => {
-        val dataSet = TextSet.array(Array(TextFeature.apply(x)))
-        // Pre-processing
-        val transformed = dataSet.setWordIndex(wordIndex)
-          .tokenize()
-          .normalize()
-          .word2idx(removeTopN = 10, maxWordsNum = param.maxWordsNum)
-          .shapeSequence(param.sequenceLength).generateSample()
-        // Predict
-        val predictSet = model.predict(transformed,
-          batchPerThread = param.partitionNum)
-        // Print result
-        predictSet.toLocal().array.map(_.getPredict)
-          .head.asInstanceOf[Tensor[Float]].toArray()
-        // Top 5 is not suit for console print
-//        predictSet.toLocal().array.map(_.getPredict)
-//          .take(5).asInstanceOf[Array[Tensor[Float]]].map(_.toArray()).toSeq
-      })
-
-      val predicts = lines.as[String].withColumn("Predict",
-        predictTextUDF(col("value")))
-
-      val query = predicts.writeStream
-        .outputMode("append")
-        .format("console")
-        .start()
-
-      query.awaitTermination()
+      lines.foreachRDD { lineRdd =>
+        if (!lineRdd.partitions.isEmpty) {
+          // RDD to TextFeature
+          println("First line " + lineRdd.top(1).apply(0))
+          val textFeature = lineRdd.map(x => TextFeature.apply(x))
+          // RDD[TextFeature] to TextSet
+          val dataSet = TextSet.rdd(textFeature)
+          // Pre-processing
+          val transformed = dataSet.setWordIndex(wordIndex)
+            .tokenize().normalize()
+            .word2idx(removeTopN = 10, maxWordsNum = param.maxWordsNum)
+            .shapeSequence(param.sequenceLength).generateSample()
+          val predictSet = model.predict(transformed,
+            batchPerThread = param.partitionNum)
+          // Print result
+          predictSet.toDistributed()
+            .rdd.take(5)
+            .map(_.getPredict.toTensor)
+            .foreach(println)
+        }
+      }
+      ssc.start()
+      ssc.awaitTermination()
     }
   }
-
 
   def word2index(word: String, wordIndex: Map[String, Int]): Int = {
     wordIndex.apply(word)
